[ANDROSDK-444] Move ForeignKeyCleaner to maintenance
